emmm (#173)
* remove babel

* prototype ok

* 初步搞定

* complete work

* rename

* fix path

* new version

* new API

* remove compat

* modify rollup config

* emmm
