--- compat/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "fre-compat",
-  "version": "1.0.0",
-  "description": "> Something that can be implemanted in userland.",
-  "main": "../dist/fre-compat.js",
-  "peerDependencies": {
-    "fre": "^2.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.10.2"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
