+++ demo/index.html
@@ -10,7 +10,7 @@
 
 <body>
   <div id="root"></div>
-  <script src="./src/resume-error.tsx"></script>
+  <script src="./src/resume-exception.tsx"></script>
 </body>
 
 </html>
