+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fre",
-  "version": "2.0.0-alpha.4",
+  "version": "2.0.0-alpha.5",
   "description": "",
   "main": "dist/fre.js",
   "unpkg": "dist/fre.js",
